@dsanders11 2020-10-19
@nornagon, both are good points. I'd forgotten that the apply patch script is being invoked with `python3` in `DEPS`, so you're right, it should in theory have already been using `-1`.

> Are you sure your measurements are accurate?

I ran it a handful of times with and without the change, and while there's some variation in the run time with this change not applied, it was 5-10 seconds.

I'll take some more rigorous data.