@dsanders11 2020-10-20
@nornagon, did more testing to confirm this is actually a no-op. Turns out there's more variability in the timings of applying the patches than I would have suspected. Tested with nothing else running on my system and doing back-to-back `e sync` operations. Originally did 10 without this change, then in a clean terminal another 10 with it, but after seeing the data continued on to 15 runs with it and went back and did another 5 runs with it.

Without change (83.394 mean, 5.51 std):

```
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 73.41 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 87.94 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 74.03 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 87.90 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 83.15 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 88.54 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 87.52 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 76.58 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 88.45 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 81.38 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 84.42 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 89.13 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 77.79 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 86.80 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 83.87 secs
```

With change (82.765 mean, 6.31 std):
```
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 75.42 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 83.46 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 84.14 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 78.85 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 89.38 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 77.72 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 88.85 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 86.32 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 75.45 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 93.95 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 84.71 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 74.66 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 88.80 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 85.59 secs
Hook 'python3 src/electron/script/apply_all_patches.py src/electron/patches/config.json' took 74.17 secs
```

So no statistical difference. Seems that there's just a fair bit of variability of run times in running that `git am` and when I tested I had some coincidental timings which made it seem like it had an effect, since the range of times is 20 seconds.